Enum scaproust::ConfigOption

pub enum ConfigOption {
    Linger(Duration),
    SendTimeout(Option<Duration>),
    SendPriority(u8),
    RecvTimeout(Option<Duration>),
    RecvPriority(u8),
    RecvMaxSize(u64),
    RetryIvl(Duration),
    RetryIvlMax(Option<Duration>),
    TcpNoDelay(bool),
    Subscribe(String),
    Unsubscribe(String),
    ReqResendIvl(Duration),
    SurveyDeadline(Duration),
}

Variants
Linger(Duration)

Specifies how long the socket should try to send pending outbound messages after drop have been called. Default value is 1 second.

SendTimeout(Option<Duration>)

See Socket::set_send_timeout.

SendPriority(u8)

See Socket::set_send_priority.

RecvTimeout(Option<Duration>)

See Socket::set_recv_timeout.

RecvPriority(u8)

See Socket::set_recv_priority.

RecvMaxSize(u64)

Maximum message size that can be received, in bytes. Zero value means that the received size is limited only by available addressable memory. Default is 1024kB.

RetryIvl(Duration)

For connection-based transports such as TCP, this option specifies how long to wait, when connection is broken before trying to re-establish it. Note that actual reconnect interval may be randomised to some extent to prevent severe reconnection storms. Default value is 0.1 second.

RetryIvlMax(Option<Duration>)

This option is to be used only in addition to ReconnectInterval option. It specifies maximum reconnection interval. On each reconnect attempt, the previous interval is doubled until ReconnectIntervalMax is reached. Value of None means that no exponential backoff is performed and reconnect interval is based only on ReconnectInterval. If RetryIvlMax is less than RetryIvl, it is ignored. Default value is None.

TcpNoDelay(bool)

See Socket::set_tcp_nodelay.

Subscribe(String)

Defined on Sub socket. Subscribes for a particular topic. A single Sub socket can handle multiple subscriptions.

Unsubscribe(String)

Defined on Sub` socket. Unsubscribes from a particular topic.

ReqResendIvl(Duration)

This option is defined on the Req socket. If a reply is not received in the specified amount of time, the request will be automatically resent. Default value is 1 minute.

SurveyDeadline(Duration)

Specifies how long to wait for responses to the survey. Once the deadline expires, receive function will return a TimedOut error and all subsequent responses to the survey will be silently dropped. Default value is 1 second.